refactor: 메서드 네이밍 수정 및 에러 핸들러 분리
