+++ src/main/java/chess/controller/ChessGameAdvice.java
@@ -1,0 +1,14 @@
+package chess.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class ChessGameAdvice {
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<String> handle(Exception exception) {
+        return ResponseEntity.badRequest().body(exception.getMessage());
+    }
+}
