+++ src/main/java/chess/controller/ChessGameController.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.controller;
 
 import chess.dao.ChessGame;
 import chess.domain.piece.property.Team;
@@ -9,7 +9,6 @@ import java.sql.SQLException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -17,25 +16,29 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
-public class WebController {
+public class ChessGameController {
 
-    private static final ChessService CHESS_SERVICE = new ChessService();
+    private final ChessService chessService;
+
+    public ChessGameController(final ChessService chessService) {
+        this.chessService = chessService;
+    }
 
     @GetMapping("/")
-    public String init(Model model) throws SQLException {
-        model.addAttribute("games", CHESS_SERVICE.getGames());
+    public String showRooms(Model model) throws SQLException {
+        model.addAttribute("games", chessService.getGames());
         return "lobby";
     }
 
     @PostMapping("/chess/new")
     public String createGame(@RequestParam String gameName) throws SQLException {
-        String gameId = CHESS_SERVICE.addChessGame(gameName);
+        String gameId = chessService.addChessGame(gameName);
         return "redirect:/chess/game/" + gameId;
     }
 
     @GetMapping("/chess/game/{id}")
-    public String temp(@PathVariable String id, Model model) {
-        ChessGameRoomInfoDTO chessGameRoomInfoDTO = CHESS_SERVICE.findGameById(id);
+    public String showChessGameRoom(@PathVariable String id, Model model) {
+        ChessGameRoomInfoDTO chessGameRoomInfoDTO = chessService.findGameById(id);
         model.addAttribute("chessGameRoom", chessGameRoomInfoDTO);
         return "game";
     }
@@ -43,7 +46,7 @@ public class WebController {
     @GetMapping("/chess/game/{id}/board")
     @ResponseBody
     public ResponseEntity<BoardDTO> createBoard(@PathVariable String id) throws SQLException {
-        ChessGame chessGame = CHESS_SERVICE.getChessGamePlayed(id);
+        ChessGame chessGame = chessService.getChessGamePlayed(id);
         return ResponseEntity.ok(new BoardDTO(chessGame));
     }
 
@@ -53,12 +56,7 @@ public class WebController {
                                               @RequestParam String target,
                                               @RequestParam String team,
                                               @PathVariable String id) throws SQLException {
-        ChessGame chessGame = CHESS_SERVICE.movePiece(id, source, target, Team.valueOf(team));
+        ChessGame chessGame = chessService.movePiece(id, source, target, Team.valueOf(team));
         return ResponseEntity.ok(new BoardDTO(chessGame));
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<String> handle(Exception exception) {
-        return ResponseEntity.badRequest().body(exception.getMessage());
     }
 }
